+++ packages/expo/src/logs/LogSerialization.ts
@@ -195,11 +195,7 @@ function _captureConsoleStackTrace(): Error {
 }
 
 function _getProjectRoot(): string | null {
-  return (
-    Constants.manifest?.developer?.projectRoot ??
-    Constants.manifest2?.extra?.expoGo?.developer?.projectRoot ??
-    null
-  );
+  return Constants.expoGoConfig?.developer?.projectRoot ?? null;
 }
 
 export default {